pyproject.toml
[tool.poetry]
name = "ewma"
version = "0.1.0"
description = "Exponentially weighted moving filter"
authors = ["Jean-David Therrien <jeandavidt@gmail.com>"]
license = "All rights reserved"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
scipy = "^1.8.0"
ipykernel = "^6.13.0"
nbformat = "^5.3.0"
ipython = "^8.2.0"
plotly = "^5.7.0"
pandas = "^1.4.2"
PyYAML = "^6.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
mypy = "^0.942"
flake8 = "^4.0.1"
pytest = "^7.1.2"
types-PyYAML = "^6.0.7"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
